re-enabled merge and increased travis_wait timeout
I decided that merge=true is a good idea, so that builds don't fail
while the (lengthy) rebuild is in progress.  For the timeout, I
don't know if 20 minutes is going to be too short or not, but it
doesn't hurt to increase it to half an hour, just in case.